Add similar products feature and enhance product-sync-service
This commit introduces the functionality to display similar products in the application. For this purpose, a new fields similar has been added in the `product.index.map.json`, along with thumb and images fields.

The product converter service has been updated to consider the new `related` field and the `similar` products. The filter for development environment is added in `product-converter.service.ts` for variants.

In addition to this, multiple changes were made in `sync-es.service.ts` such as removal of redundant `httpService` and changes in constructor.

A new exception `CouldNotCreateIndexException` is introduced. A check for ElasticSearchIndexes presence on bootstrap is implemented in `eshop.module.ts`

The changes ensure the sync service is more robust, and similar products enhance UX.