"Add ModelManager Module and add models.js to client configs"
Implemented ModelManager module and services that allows to synchronize configurable models and their fields. The new model-manager provides an API endpoint to update model fields dynamically. It also validates the structure of the fields before applying the changes.

Additionally, an update script is provided to add a 'models.js' file to client configs. This file serves the purpose of overriding or adding new models. These changes provide greater flexibility for managing models in the application as well as improves the client configuration system.